requirements-test.txt
# Testing dependencies
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.24.1

# API testing
fastapi==0.103.1
uvicorn[standard]==0.23.2

# Required for tests
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
numpy==1.24.3
pydantic==2.3.0
python-multipart==0.0.6
